requirements.txt
asgiref>=3.2.7
autopep8>=1.5.2
boto>=2.49.0
certifi>=2020.4.5.1
chardet>=3.0.4
Django==2.2.12
django-crispy-forms>=1.9.1
django-haystack>=2.8.1
django-js-asset>=1.2.2
django-machina>=1.1.1
django-mptt>=0.11.0
django-ses-gateway>=0.1.1
django-simple-pagination>=1.3
django-storages>=1.9.1
django-widget-tweaks>=1.4.8
idna>=2.9
lxml>=4.5.1
markdown2>=2.3.9
microurl>=0.1.1
Pillow>=7.1.2
pycodestyle>=2.6.0
python-http-client>=3.2.7
pytz>=2020.1
requests>=2.23.0
sendgrid>=6.3.1
six>=1.15.0
sorl-thumbnail>=12.6.3
sqlparse>=0.3.1
urllib3>=1.25.9
